requirements.txt
aiogram
aiogram_calendar
google-generativeai